+++ client/src/components/App.js
@@ -9,7 +9,7 @@ import Dashboard from './pages/Dashboard'
 import Statistics from './pages/Statistics'
 import Logs from './pages/Logs'
 import InsertUser from './pages/InsertUserPage'
-import ModifyUser from './pages/ModifyUser'
+import ModifyUser from './pages/ModifyUserPage'
 import DeleteUser from './pages/DeleteUserPage'
 import NotFound from './pages/NotFound'
